Updated open office and closed to be unique
Based on some research I adjusted the density and EPD down for closed
offices and up for private offices.  I don't know that my values are
correct, but they are better than they were when they were the same as
each other. In the end, this is a starting point and modelers can adjust
based on their specific building design.