old spreadsheet on ref bldg. constructions
I did have one what had R values by vintage and climate zone. If I can't
find that, all of that information is in the ref building SVN, which I'm
checking out now, and of course it is in our current templates. If we
want to work with pre-made constructions an Ngrid script parses the
const object names to pick the right one.