Fixed missing SchRulset and Sch Set fields
Odd, the source file for this was fine, but the schedule type limits
object for fractional schedules from the source file isn't in the master
file. Maybe because of this all of the fields for those schedule objects
were emptied out.

In a few cases even Schedule set objects were emptied out. If I missed
anything data can be copied back from Jan2014_OfficeSchedules.osm

This is the UUID I had been using for schedule type limits for
fractional schedules
{d9b4a865-0dd0-4333-8fdb-88a55fb07f16}

This is the valid uuid for a fraction type limits object in the master
file. This is what I used.
{0b8a95eb-9aa8-43a9-9736-998a36816f12}