add dedicated method for PRM chilled water plant creation
This removes the dependency on the prototype methods for CHW plant
creation. This creates CHW  and CW loops more inline with App. G. Should
hopefully reduce execution time as one less sizing run has to be
performed because this method uses the ideal cooling loads for the
primary zones to predict needed cooling capacity.