Fixing expected values for Models 20 and 21
Model 20 is a large office, and has 4 modeled floors and 20 modeled thermal zones.
Expected values for number of air loops and number of heating coils has been updated to respectively 4 (one per floor) and 23 (one per floor and one per thermal zones).

Model 21 is a warehouse with only heating and should therefore has a baseline model with System 10 (the test was previously checking for System 6)