Autosizing minmium outdoor airflow rate
The minimum outdoor airflow rate was never set to be autosized (the default value is 0) for PSZ-AC systems. One of the consequences is that no ERVs were created because, for most standards, the ERV requirement depends on the percentage of outdoor air flow rate to design flow rate which was always 0.