Fix/Retail standalone ventilation 90.1-2004
The 90.1-2004 standalone retail prototype could not be created because of a severe error in the sizing run. No outdoor air was being assigned to the core retail space. The proposed changes added a line in the Space Types tab of the Google Sheet for this space (it was already defined for all the other 90.1 standards except for 90.1-2004). The test didn't catch this because it only creates the 2007 and 2013 version of 90.1 prototypes.