change radiant operation to apply to all environment periods except warmup
The CurrentEnvironment variable is dependent on the number of design days in the simulation, so a hard integer check, e.g. CurrentEnvironment == 3, will not always correspond to the annual weather run period.  Using the WarmupFlag instead to define warmup days and setting radiant operation to follow the annual weather run logic for all other environment periods including design days.