change design spec oa method to retain values for DCV
If a zone or system did not require DCV, the values in the design specification outdoor air object were all converted to flow per area.  This means that DCV cannot be applied to the model in a future measure, as that information is lost.  This refactor changes the conversion to a flow per zone and sets the oa method to Flow/Zone to retain per person values for future DCV application.