change instances of addBranchForZone to multiAddBranchForZone
The addBranchForZone method will disconnect existing air terminals.  multiAddBranchForZone does not and allows multiple air loops and diffusers to serve the same zone.
This is a rare case that occurs when a ventilation system (DOAS, VAV) is paired with a zone system that uses the AirLoopHVACUnitarySystem object (minisplits, evaporative coolers, etc.).