check that adjacent surface exists for plenums
the tall building methods delete some stories dynamically depending on the ratio of floor types (e.g. one floor might be retail).  Deleting a space will not chance the surface adjacencies in adjacent spaces, which will still show surface adjacencies.  This causes an error in the add constructions methods, which try to set plenum constructions for the ceiling if the adjacent space is a plenum.  As there was no adjacent space, the get method errored out.