remove construction changes from json
- remove change_construction_properties_in_model
- remove apply_changes_to_surface_construction

apply_changes_to_surface_construction is only used by change_construction_properties_in_model to change u-values and window shgc/tvis for simple glazing based on a json file. This may be a useful method, but there are clearer ways to implement it. A revised version of these methods may be added back in the future.