allow multiple radiant slab boundary conditions
The radiant slab construction was only checking for 'Ground' and 'Outdoors' conditions, and treating everything else and interior. Then it would try to find an adjacent surface. That did not exist for models with the 'GroundFCfactorMethod' boundary condition. Adjust to allow multiple radiant slab ground boundary conditions, and check if an adjacent surface exists before getting it.