remove school office secondary daylighting control
The check for adding secondary daylighting used to be:
`if areas['secondary_sidelighted_area'] == 0.0`
This branch changed it to:
`if areas['secondary_sidelighted_area'] < 0.01`
which triggered a regression error in primary school offices. These should not have secondary daylighting (the zone area fraction served is so small, it is recorded as 0 in the thermal zone object.
Remove secondary daylighting, as it is not required per code.
This is another instance where rubocop caught and suggested a fix for float comparisons == 0.0 not operating as intended.