add perimeter tolerance check
Add a perimeter tolerance for double loaded corridors. If there is a double loaded corridor, the perimeter distance will change to keep the area ratio between the primary space type and circulation space type. If the resulting custom perimeter depth is close to the specified perimeter zone depth (usually 15ft), there may be a mismatch in surface area in surface matching. This fix uses the perimeter zone depth instead of the custom value if it is within 1 millimeter of the perimeter zone depth.