Nrcan merge master (#1995)
* Splits up the tests in test_appendix_g_prm

The test_appendix_g_prm.rb test is a collection of tests. The way the parallelization works means this tests runs as one unit and takes 11 hours to finish. Splitting it up into separate unit tests decreases run time on the CI.

* rename test prm test classes

don't define all as AppendixGPRMTests

* Revert "rename test prm test classes"

This reverts commit 5fa3b64cd553fa02c009f168d41d3f24f5ba5a60.

* coil dx terminal application fix

The efficiency lookup tables require that the application field not be 'nil'. The method to determine it is fixed on a few ASHRAE templates; all other templates get nil. Change to an else statement so other templates get this field assigned.

* add comstock pthp test to create_typical tests

* include capacity in cannot find efficiency warnings

* Add additional coil types to DX heating type lookup

Several DX cooling coils that were part of unitary systems were returning a warning that the matched the criteria for multiple efficiency lookups. In this case, it was unitary systems with DX cooling coils and hot water heating coils. Add support for that type of coil combination.

* Updating covert_arg_to_bool to work as intended

* Updating some files which were eroneously updated.

---------

Co-authored-by: Matthew Dahlhausen <matthew.dahlhausen@gmail.com>
Co-authored-by: Landan Taylor <128543626+LandanT@users.noreply.github.com>