fix coil eff application to necb models
The NECB PTAC / PTHP efficiency data doesn't have the 'application'  field, causing the  lookup to return nil, and the efficiency application to return false. Since the vars hash gets successively passed, further methods failing when trying to set variables on the now false hash.
Fixes are to return the original hash instead of false in the apply method. And to add a check to only apply the application search field for 90.1 templates.