Including additional changes from master and some fixes (#2001)
* data_update b8458b8

* Splits up the tests in test_appendix_g_prm

The test_appendix_g_prm.rb test is a collection of tests. The way the parallelization works means this tests runs as one unit and takes 11 hours to finish. Splitting it up into separate unit tests decreases run time on the CI.

* rename test prm test classes

don't define all as AppendixGPRMTests

* change water temperatures to mixed temperatures

- Change water use temperatures to be the mixed water temperature. This is necessary because the faucet temps should be lower to account for heat loss in pipes and mixing. The 140F water use temps the same as the loop and water heater temps were causing issues with sizing. The loop rarely hit the target temperature.
- Mixed water temperatures changed from 140F to 110F for most use cases, 120F in kitchens for general use. Keep 180F for dishwashers.
- The prior flow rates and schedules seemed to be mixed water use, not hot water use. As an example, a 20,000 ft2 office with 1000 ft2 of bathrooms would have 5 EFLH * 0.02 gph per ft2 * 1000 ft2 = 100 gallons mixed water use, whereas 100 gallons of 140F would be to high. The overall values will need to be checked and updated.

* Revert "rename test prm test classes"

This reverts commit 5fa3b64cd553fa02c009f168d41d3f24f5ba5a60.

* coil dx terminal application fix

The efficiency lookup tables require that the application field not be 'nil'. The method to determine it is fixed on a few ASHRAE templates; all other templates get nil. Change to an else statement so other templates get this field assigned.

* add comstock pthp test to create_typical tests

* include capacity in cannot find efficiency warnings

* Add additional coil types to DX heating type lookup

Several DX cooling coils that were part of unitary systems were returning a warning that the matched the criteria for multiple efficiency lookups. In this case, it was unitary systems with DX cooling coils and hot water heating coils. Add support for that type of coil combination.

* set water heater sizing properties

- set minimum loop temperatures of 125F for service water heating loops for legionella risk
- change the sizing loop delta temperature to larger values
- set water heater tank recovery times depending on fuel, and set sizing to PeakDraw

* Updating covert_arg_to_bool to work as intended

* Updating some files which were eroneously updated.

* fix coil eff application to necb models

The NECB PTAC / PTHP efficiency data doesn't have the 'application'  field, causing the  lookup to return nil, and the efficiency application to return false. Since the vars hash gets successively passed, further methods failing when trying to set variables on the now false hash.
Fixes are to return the original hash instead of false in the apply method. And to add a check to only apply the application search field for 90.1 templates.

* change necb models to new minimum loop temperature

* update expected regression results

* fix comstock pthp efficiency lookup

90.1-2004 and 90.1-2007 ASHRAE heat pump heating efficiencies use 'New Construction' and 'Replacements' as the application type, rather than 'Standard Size'. When the template includes ComStock , it uses the ASHRAE 90.1 databases but didn't strictly match the template type, so the efficiency lookup would fail.

* change swh loop temperature difference to 5K

change the service hot water loop design temperature difference back to 5 Kelvin (9 Rankine) from prior values. The prior changes were causing very low temperatures on booster hot water loops. And more typical design differences are 5-10R.

* Replacing openstudio-standards require statment with require_relative path to openstudio-standards.rb.

* updates regression models

* Update version.rb

* Updating test_run_costing_tests.rb to use initial set of building types.

---------

Co-authored-by: Xuechen (Jerry) Lei <xuechen.lei@pnnl.gov>
Co-authored-by: Matthew Dahlhausen <matthew.dahlhausen@gmail.com>
Co-authored-by: Landan Taylor <128543626+LandanT@users.noreply.github.com>