method and data for thermostat schedule lookup
Use the ComStock templates to generate a thermostat_schedule_lookup.json.
The thermostat schedules will be attached to space types, with variation by building type.
The ComStock template will disappear soon; this commit is a log of how the thermostat schedule_lookup was generated, using the templates as they existed in openstudio-standards. (0.8.4, with edits to support incomplete space types. Possibly 0.8.5 - see commits around this time.).